Make the duty cycle a fraction so it doesn't have to worry about pwm rate. Also caugth a few little bugs and enable the PWM outputs.

+++ flight/PiOS/inc/pios_esc.h
@@ -36,7 +36,7 @@
 void PIOS_ESC_Off();
 void PIOS_ESC_Arm();
 void PIOS_ESC_NextState();
-void PIOS_ESC_SetDutyCycle(uint16_t duty_cycle);
+void PIOS_ESC_SetDutyCycle(float duty_cycle);
 void PIOS_ESC_SetState(uint8_t new_state);
 
 #endif /* PIOS_DELAY_H */
